+++ RELEASES.md
@@ -38,6 +38,8 @@ See also GitHub releases https://github.com/lucmoreau/ProvToolbox/releases
 * service-templates-core: first version
 * prov-template-archetype: first version
 * prov-template-compiler: changed Locations to better manager packages of generated code
+* prov-template-compiler: added generated method isCompositeOf in Builder
+* prov-template-archetype: support for compulsory properties in forms.html for composites
 
 `2.2.1`
 * service-metrics: a service to gather metrics on provenance, using MetricsCalculator.java